Simplify
When I created this project with Android Studio, I chose "Fullscreen
Activity" in the "Add an Activity to Mobile" page of the project
creation wizard.  That was a mistake.  I therefore created a new
project, chose "Empty Activity" this time, and merged the new project's
files into this project as well as possible.

Also see [1].

[1]: https://developer.android.com/training/basics/firstapp/creating-project