Fix endless loop of restarting the camera app
The reason for this behavior probably was that my app's process was
typically killed to free memory after starting the camera app, and
restarted when leaving the camera app.  Then my app would start the
camera app again.  The fix is to save and restore some state.

See [1].  For example:

> If an activity is completely obscured by another activity, it is
> stopped.  It still retains all state and member information, however,
> it is no longer visible to the user so its window is hidden and it
> will often be killed by the system when memory is needed elsewhere."

[1]: https://developer.android.com/reference/android/app/Activity