Fix showing text and spinny thing simultaneously
It's possible that the `TextView` and the `ProgressBar` are both shown
at the same time.  They occupy the same position, so this isn't supposed
to happen.

It can happen when we take a new photo while an `AsyncTask` is already
running.  The value of the `TextView` is only set to `null` after taking
a photo, so if the `AsyncTask` doesn't complete before the `TextView` is
set to `null`, whatever the `AsyncTask` sets the `TextView` to later
will not be cleared, but the new `AsyncTask` starts and the spinny thing
is shown.

Note that "`AsyncTask`s [are] executed serially on a single background
thread." [1] (multiple threads were only used in a few Android versions
a long, long time ago.)

Anyway, fix stuff by making sure the `score` property of the `ViewModel`
is set to `null` when an obsolete `AsyncTask` completes (this results in
the `TextView` being cleared).

[1]: https://developer.android.com/reference/kotlin/android/os/AsyncTask