Generate icons inside circles
Generate icons based on `icon-for-circles.png` using Android Studio
because somebody at Google decided that everything is a circle in new
Android versions or something.

The only difference between `icon.png` and `icon-for-circles.png` is
that I changed the padding.  The new padding is about right for having
the icon centered inside the circle based on its centroid.